# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Go Email Service Build with Go-Fiber and Go-Imap",
        "title": "Go Email Service API",
        "contact": {
            "name": "Bipin Maharjan",
            "email": "bipinmhr10@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login Body",
                        "name": "RequestBody",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.LoginBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "description": "Logout API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/mailbox": {
            "get": {
                "description": "get List of Mailboxes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mailbox"
                ],
                "summary": "Get list of Mailboxes",
                "responses": {}
            }
        },
        "/message/preview": {
            "get": {
                "description": "Get All Preview Message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Get All Preview Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailbox Name",
                        "name": "mailboxName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 1,
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": 20,
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "controller.LoginBody": {
            "type": "object",
            "required": [
                "host",
                "password",
                "port",
                "username"
            ],
            "properties": {
                "host": {
                    "type": "string",
                    "example": "outlook.office365.com"
                },
                "password": {
                    "type": "string",
                    "format": "string",
                    "example": "password1122"
                },
                "port": {
                    "type": "string",
                    "example": "993"
                },
                "username": {
                    "type": "string",
                    "format": "string",
                    "example": "test@test.com"
                }
            }
        }
    }
}